on progress: Redefine database for relationship between entity

+++ app/src/main/java/com/example/academictaskorganizer_skripsi/view/SubjectAdapter.kt
@@ -103,19 +103,19 @@ class SubjectDiffCallback : DiffUtil.ItemCallback<SubjectDataItem>() {
 
 }
 
-class SubjectListener(val clickListener: (subjectId: Long) -> Unit)
+class SubjectListener(val clickListener: (subjectId: Int) -> Unit)
 {
     fun onClick(subject: Subject) = clickListener(subject.subjectId)
 }
 
 sealed class SubjectDataItem {
-    abstract val id: Long
+    abstract val id: Int
     data class SubjectItem(val subject: Subject): SubjectDataItem(){
         override val id = subject.subjectId
     }
 
     object Header: SubjectDataItem(){
-        override val id = Long.MIN_VALUE
+        override val id = Int.MIN_VALUE
     }
